3D Models (Platforms)
Heyyyyy here's the 3D models for the fruitger aero level + the cat cafe level! Also I'm not sure if we want a more smooth aesthetic or more blocky- it's nothing tough to switch between them at this point- just let me know. Also the cat pudding is just like an extra fun thing that I saw on tiktok that maybe we can add as a jump pad if we have time :p Some of these items will have animations in the future and ofc none of them have textures